NH - MF-217 - Identifier component refactor/directory setup
