+++ package.json
@@ -41,6 +41,7 @@
     "@openmrs/react-root-decorator": "^2.0.0",
     "@testing-library/react": "^9.1.3",
     "@types/jest": "^24.0.18",
+    "@types/lodash": "^4.14.161",
     "@types/react": "^16.9.2",
     "@types/react-dom": "^16.9.0",
     "@types/react-router": "^5.1.7",
@@ -81,8 +82,8 @@
   "dependencies": {
     "@openmrs/esm-api": "^1.2.2",
     "@openmrs/esm-module-config": "^1.2.0",
-    "formik": "^2.1.5",
     "@openmrs/esm-styleguide": "^1.3.0",
+    "formik": "^2.1.5",
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
     "yup": "^0.29.1"
