+++ src/patient-registration/patient-registration.resource.tsx
@@ -76,5 +76,4 @@ export async function getSecondaryIdentifierTypes(abortController: AbortControll
   }
 }
 
-export const uuidIdentifier = '05a29f94-c0ed-11e2-94be-8c13b969e334';
 export const uuidTelephoneNumber = '14d4f066-15f5-102d-96e4-000c29c2a5d7';
