+++ src/patient-registration/section/identifier/identifier-section.component.tsx
@@ -1,7 +1,7 @@
-import React, { useEffect } from 'react';
-import { IdentifierInput } from '../input/custom-input/identifier-input.component';
-import { PatientIdentifierType } from '../patient-registration-helper';
-import styles from './section.css';
+import React from 'react';
+import { IdentifierInput } from '../../input/custom-input/identifier/identifier-input.component';
+import { PatientIdentifierType } from '../../patient-registration-helper';
+import styles from './../section.css';
 
 export const IdentifierSection: React.FC<{ identifierTypes: PatientIdentifierType[] }> = ({ identifierTypes }) => {
   return (
